.history/requirements_20251108202437.txt
pandas
numpy
nltk
regex
scikit-learn
google-api-python-client
googleapiclient
streamlit
plotly
matplotlib
wordcloud
python-dotenv
tqdm
